+++ src/main/kotlin/ru/yandex/money/gradle/plugins/moira/trigger/MoiraTriggerPlugin.kt
@@ -5,7 +5,10 @@ import org.gradle.api.Project
 import org.gradle.api.artifacts.Configuration
 import org.gradle.api.plugins.JavaBasePlugin
 import org.gradle.api.plugins.JavaPluginConvention
+import org.gradle.api.tasks.Copy
 import java.io.File
+import java.nio.file.Paths
+import kotlin.streams.toList
 
 /**
  * Plugin for compiling and uploading Moira triggers.
@@ -18,44 +21,86 @@ open class MoiraTriggerPlugin : Plugin<Project> {
     override fun apply(target: Project) {
         val extension = target.extensions.create(EXTENSION_NAME, MoiraTriggerExtension::class.java)
         target.pluginManager.apply(JavaBasePlugin::class.java)
-        val configuration = target.configureSourceSets(extension)
+        val dirConfiguration = target.configureDirSourceSets(extension)
+        val artifactConfiguration = target.configureArtifactSourceSets()
+
+        val triggersConfiguration = target.configurations
+                .maybeCreate("${ARTIFACT_TRIGGERS_CONFIGURATION_NAME}Compile")
+
         target.afterEvaluate { project ->
-            project.createUploadMoiraTriggersTask(configuration, extension)
-            project.createCollectMoiraTriggersTask(configuration, extension)
+            artifactConfiguration.extendsFrom(triggersConfiguration)
+            project.createUploadMoiraTriggersTask(dirConfiguration, artifactConfiguration, extension)
+            project.createCollectMoiraTriggersTask(dirConfiguration, artifactConfiguration, extension)
+            project.createExtractTriggersTask(triggersConfiguration)
         }
     }
 
-    private fun Project.configureSourceSets(extension: MoiraTriggerExtension): Configuration {
+    private fun Project.configureDirSourceSets(extension: MoiraTriggerExtension): Configuration {
         val sourceSets = convention.getPlugin(JavaPluginConvention::class.java).sourceSets
-        val sourceSet = sourceSets.create(SOURCE_SET_NAME)
+        val sourceSet = sourceSets.create(DIR_SOURCE_SET_NAME)
         sourceSet.java.srcDir(File(extension.dir))
 
-        dependencies.add(
-            CONFIGURATION_NAME,
-            "org.jetbrains.kotlin:kotlin-stdlib:${KotlinVersion.CURRENT}"
-        )
-        dependencies.add(
-            CONFIGURATION_NAME,
-            "org.jetbrains.kotlin:kotlin-reflect:${KotlinVersion.CURRENT}"
-        )
-
-        return configurations.getByName(CONFIGURATION_NAME)
+        return addKotlinDependencies("${DIR_SOURCE_SET_NAME}Compile")
     }
 
-    private fun Project.createUploadMoiraTriggersTask(configuration: Configuration, extension: MoiraTriggerExtension) {
+    private fun Project.configureArtifactSourceSets(): Configuration {
+        val sourceSets = convention.getPlugin(JavaPluginConvention::class.java).sourceSets
+        val sourceSet = sourceSets.create(ARTIFACT_SOURCE_SET_NAME)
+        sourceSet.java.srcDir(Paths.get(buildDir.toString(), TRIGGERS_FROM_ARTIFACT_DIR).toString())
+
+        return addKotlinDependencies("${ARTIFACT_SOURCE_SET_NAME}Compile")
+    }
+
+    private fun Project.addKotlinDependencies(configurationName: String): Configuration {
+        dependencies.add(
+                configurationName,
+                "org.jetbrains.kotlin:kotlin-stdlib:${KotlinVersion.CURRENT}"
+        )
+        dependencies.add(
+                configurationName,
+                "org.jetbrains.kotlin:kotlin-reflect:${KotlinVersion.CURRENT}"
+        )
+        return configurations.getByName(configurationName)
+    }
+
+    private fun Project.createUploadMoiraTriggersTask(
+        dirConfiguration: Configuration,
+        artifactConfiguration: Configuration,
+        extension: MoiraTriggerExtension
+    ) {
         val task = tasks.create(UPLOAD_MOIRA_TRIGGERS_TASK_NAME, UploadMoiraTriggersTask::class.java)
         task.group = "other"
         task.description = "Upload Moira triggers"
-        task.configuration = configuration
+        task.dirConfiguration = dirConfiguration
+        task.artifactConfiguration = artifactConfiguration
         task.extension = extension
     }
 
-    private fun Project.createCollectMoiraTriggersTask(configuration: Configuration, extension: MoiraTriggerExtension) {
+    private fun Project.createCollectMoiraTriggersTask(
+        dirConfiguration: Configuration,
+        artifactConfiguration: Configuration,
+        extension: MoiraTriggerExtension
+    ) {
         val task = tasks.create(COLLECT_MOIRA_TRIGGERS_TASK_NAME, CollectMoiraTriggersTask::class.java)
         task.group = "other"
         task.description = "Collect Moira triggers without uploading"
-        task.configuration = configuration
+        task.dirConfiguration = dirConfiguration
+        task.artifactConfiguration = artifactConfiguration
         task.extension = extension
+    }
+
+    private fun Project.createExtractTriggersTask(triggersConfiguration: Configuration) {
+        val task = tasks.create("extractMoiraTriggers", Copy::class.java)
+
+        val files = triggersConfiguration.files.stream()
+                .map { file -> zipTree(file) }
+                .toList()
+
+        task.from(files)
+        task.into(Paths.get(buildDir.toString(), TRIGGERS_FROM_ARTIFACT_DIR).toString())
+
+        tasks.getByName(UPLOAD_MOIRA_TRIGGERS_TASK_NAME).dependsOn(task)
+        tasks.getByName(COLLECT_MOIRA_TRIGGERS_TASK_NAME).dependsOn(task)
     }
 
     companion object {
@@ -66,14 +111,24 @@ open class MoiraTriggerPlugin : Plugin<Project> {
         private const val EXTENSION_NAME = "moira"
 
         /**
-         * Plugin source set name.
+         * Plugin source set name (from directory)
          */
-        private const val SOURCE_SET_NAME = "moira"
+        private const val DIR_SOURCE_SET_NAME = "moiraFromDir"
 
         /**
-         * Plugin configuration name.
+         * Plugin source set name (from artifact)
          */
-        private const val CONFIGURATION_NAME = "moiraCompile"
+        private const val ARTIFACT_SOURCE_SET_NAME = "moiraFromArtifact"
+
+        /**
+         * Directory, where triggers from libraries will be saved
+         */
+        const val TRIGGERS_FROM_ARTIFACT_DIR = "moira"
+
+        /**
+         * Configuration, which contains artifacts with triggers
+         */
+        private const val ARTIFACT_TRIGGERS_CONFIGURATION_NAME = "moiraTriggers"
 
         /**
          * The name of task that uploads triggers to Moira via HTTP calls.