+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=MAJOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+1. Добавлен sourceSet moiraTriggersCompile, куда можно добавлять артефакты, содержащие описания триггеров.
+2. SourceSet разделен на moiraFromArtifact и moiraFromDir для возможности добавления moira-dsl разных версий.
 ### NEXT_VERSION_DESCRIPTION_END
 ## [1.0.0]() (04-07-2019)